Changed to only replace misisng links.

+++ fixLinks.sh
@@ -12,9 +12,13 @@ echo "Changing to the $dir directory"
 cd $dir || exit
 echo "...done"
 
-# TODO: Fix behavor when fixing a symlink to a directory dotfile
+echo "Fixing any missing links."
+
 while read -r -d '' file; do
-    echo "Fixing symlink to $file in home directory."
-    ln -sf "$dir/$file" "$HOME/$file"
+    if [ ! -e "$HOME/$file" ]; then
+        echo "Fixing symlink to $file in home directory."
+        ln -sf "$dir/$file" "$HOME/$file"
+    fi
 done < <(find . -name ".*" ! -name "." ! -name ".git" ! -name ".DS_Store" -print0 | sed "s/\.\///g")
 
+echo "...done"
